# Optimized std.datetime.date.Date.fromISOExtString